Renamed the bot and related files, and shifted some parts of the main function out
